// src/app/interfaces/ricky-and-morty.interface.ts
export interface RickyAndMorty {
    info: Info;
    results: Personaje[];
}

export interface Info {
    count: number;
    pages: number;
    next: string;
    prev: null;
}

export interface Personaje {
    id: number;
    name: string;
    status: Status;
    species: Species;
    type: string;
    gender: Gender;
    origin: Location;
    location: Location;
    image: string;
    episode: string[];
    url: string;
    created: Date;
}

export enum Gender {
    Female = "Female",
    Male = "Male",
    Unknown = "unknown",
}

export interface Location {
    name: string;
    url: string;
}

export enum Species {
    Alien = "Alien",
    Human = "Human",
}

export enum Status {
    Alive = "Alive",
    Dead = "Dead",
    Unknown = "unknown",
}